+++ UnrealEngine/NuitrackPlugin/Source/NuitrackModule/NuitrackModule.cpp
@@ -1,4 +1,0 @@
-#include "NuitrackModule.h"
-#include <Runtime/Core/Public/Modules/ModuleManager.h>
-
-IMPLEMENT_GAME_MODULE(FDefaultGameModuleImpl, NuitrackModule)