+++ UnrealEngine/NuitrackPluginForUE4/Source/NuitrackModule/include/nuitrack/capi/NuitrackModule_CAPI.h
@@ -1,0 +1,8 @@
+// Copyright 3DiVi 2024, Inc. All Rights Reserved.
+
+#ifndef NUITRACK_NUITRACKMODULE_CAPI_H_
+#define NUITRACK_NUITRACKMODULE_CAPI_H_
+
+#include "nuitrack/modules/Module.h"
+
+#endif /* NUITRACK_NUITRACKMODULE_CAPI_H_ */