+++ UnrealEngine/NuitrackPluginForUE4/Source/NuitrackModule/include/nuitrack/types/Color3.h
@@ -1,0 +1,25 @@
+// Copyright 3DiVi 2024, Inc. All Rights Reserved.
+
+#ifndef NUITRACK_COLOR3_H_
+#define NUITRACK_COLOR3_H_
+
+namespace tdv
+{
+namespace nuitrack
+{
+
+extern "C" {
+
+struct Color3
+{
+	uint8_t blue;
+	uint8_t green;
+	uint8_t red;
+};
+
+}
+
+}
+}
+
+#endif /* NUITRACK_COLOR3_H_ */