+++ UnrealEngine/NuitrackPluginForUE4/Source/NuitrackPlugin/NuitrackPlugin.Build.cs
@@ -1,0 +1,13 @@
+using UnrealBuildTool;
+
+public class NuitrackPlugin : ModuleRules
+{
+	public NuitrackPlugin(ReadOnlyTargetRules Target) : base(Target)
+	{
+		PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
+		
+		PublicDependencyModuleNames.AddRange(new string[] { "Core" });
+		
+		PrivateDependencyModuleNames.AddRange(new string[] { "CoreUObject", "Engine", "Slate", "SlateCore", "NuitrackModule" });
+    }
+}